Run a Stan model and return only the summaries.

Not a user-side function. Do not invoke directly.

tar_stan_gq_rep_summary_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  copy_data,
  variables,
  summaries,
  summary_args
)

Arguments

data
Same as for the $sample() method.

fitted_params
Symbol, name of a CmdStanFit object computed in a previous target: for example, the *_mcmc_* target from tar_stan_mcmc(). Must be a subclass that $generate_quantities() can accept as fitted_params.

compile
Do compilation? The default is TRUE. If FALSE compilation can be done later via the $compile() method.

quiet
(logical) Should the verbose output from CmdStan during compilation be suppressed? The default is TRUE, but if you encounter an error we recommend trying again with quiet=FALSE to see more of the output.

dir
(string) The path to the directory in which to store the CmdStan executable (or .hpp file if using $save_hpp_file()). The default is the same location as the Stan program.

include_paths
(character vector) Paths to directories where Stan should look for files specified in #include directives in the Stan program.

cpp_options
(list) Any makefile options to be used when compiling the model (STAN_THREADS, STAN_MPI, STAN_OPENCL, etc.). Anything you would otherwise write in the make/local file.

stanc_options
(list) Any Stan-to-C++ transpiler options to be used when compiling the model. See the Examples section below as well as the stanc chapter of the CmdStan Guide for more details on available options: https://mc-stan.org/docs/cmdstan-guide/stanc.html.

force_recompile
(logical) Should the model be recompiled even if was not modified since last compiled. The default is FALSE.

seed
Same as for the $sample() method.

output_dir
Same as for the $sample() method.

sig_figs
Same as for the $sample() method.

parallel_chains
Same as for the $sample() method.

threads_per_chain
Same as for the $sample() method.

variables
(character vector) The variables to include.

summaries
Optional list of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
Optional list of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

Value

A data frame of posterior summaries.