One MCMC with multiple outputs.

Targets to run a Stan model once with MCMC and save multiple outputs.

tar_stan_mcmc(
  name,
  file,
  data = list(),
  compile = c("original", "copy"),
  quiet = TRUE,
  dir = NULL,
  include_paths = NULL,
  cpp_options = list(),
  stanc_options = list(),
  force_recompile = FALSE,
  seed = NULL,
  refresh = NULL,
  init = NULL,
  save_latent_dynamics = FALSE,
  output_dir = NULL,
  chains = 4,
  parallel_chains = getOption("mc.cores", 1),
  chain_ids = seq_len(chains),
  threads_per_chain = NULL,
  iter_warmup = NULL,
  iter_sampling = NULL,
  save_warmup = FALSE,
  thin = NULL,
  max_treedepth = NULL,
  adapt_engaged = TRUE,
  adapt_delta = NULL,
  step_size = NULL,
  metric = NULL,
  metric_file = NULL,
  inv_metric = NULL,
  init_buffer = NULL,
  term_buffer = NULL,
  window = NULL,
  fixed_param = FALSE,
  sig_figs = NULL,
  validate_csv = TRUE,
  show_messages = TRUE,
  variables = NULL,
  inc_warmup = FALSE,
  summaries = list(),
  summary_args = list(),
  draws = TRUE,
  diagnostics = TRUE,
  summary = TRUE,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Symbol, base name for the collection of targets. The name itself will be applied to the fit object itself, and there will be suffixes for various supporting targets.

file
Code to generate the stan_file argument of $compile(). Could just be a literal path to a Stan model file or the name of an upstream target defined by tar_stan_compile().

data
Code to generate the data argument of $sample().

compile
Do compilation? The default is TRUE. If FALSE compilation can be done later via the $compile() method.

quiet
quiet argument to $compile().

dir
dir argument to $compile().

include_paths
include_paths argument to $compile().

cpp_options
cpp_options argument to $compile().

stanc_options
stanc_options argument to $compile().

force_recompile
force_recompile argument to $compile().

seed
seed argument to $sample().

refresh
refresh argument to $sample().

init
init argument to $sample().

save_latent_dynamics
save_latent_dynamics argument to $sample().

output_dir
output_dir argument to $sample().

chains
chains argument to $sample().

parallel_chains
parallel_chains argument to $sample().

chain_ids
chain_ids argument to $sample().

threads_per_chain
threads_per_chain argument to $sample().

iter_warmup
iter_warmup argument to $sample().

iter_sampling
iter_sampling argument to $sample().

save_warmup
save_warmup argument to $sample().

thin
thin argument to $sample().

max_treedepth
max_treedepth argument to $sample().

adapt_engaged
adapt_engaged argument to $sample().

adapt_delta
adapt_delta argument to $sample().

step_size
step_size argument to $sample().

metric
metric argument to $sample().

metric_file
metric_file argument to $sample().

inv_metric
inv_metric argument to $sample().

init_buffer
init_buffer argument to $sample().

term_buffer
term_buffer argument to $sample().

window
window argument to $sample().

fixed_param
fixed_param argument to $sample().

sig_figs
sig_figs argument to $sample().

validate_csv
validate_csv argument to $sample().

show_messages
show_messages argument to $sample().

variables
variables argument to $draws() and $summary() on the CmdStanMCMC object.

inc_warmup
inc_warmup argument to $draws().

summaries
Optional list of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
Optional list of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

draws
Logical, whether to create a target for posterior draws. Saves posterior::as_draws_df(fit$draws()) to a compressed tibble. Convenient, but duplicates storage.

diagnostics
Logical, whether to create a target for posterior::as_draws_df(fit$sampler_diagnostics()). Saves posterior::as_draws_df(fit$draws()) to a compressed tibble. Convenient, but duplicates storage.

summary
Logical, whether to create a target for fit$summary().

tidy_eval
Logical, whether to enable tidy evaluation when interpreting command and pattern. If TRUE, you can use the "bang-bang" operator !! to programmatically insert the values of global objects.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier. Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle()).

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

tar_stan_mcmc(name = x, ...) returns a list of targets::tar_target() objects:

x_file: reproducibly track the Stan model file.

x_lines: contents of the Stan model file. Omitted if compile = "original".

x_data: data for the MCMC.

x_mcmc: CmdStanMCMC object with all the MCMC results.

x_draws: tidy data frame of MCMC draws. Omitted if draws = FALSE.

x_summary: tidy data frame of MCMC summaries. Omitted if summary = FALSE.

x_diagnostics: tidy data frame of MCMC sampler diagnostics. Omitted if diagnostics = FALSE.

Details

Most of the arguments are passed to the $compile(), $sample(), and $summary() methods of the CmdStanModel class. If you previously compiled the model in an upstream tar_stan_compile() target, then the model should not recompile.

Examples

# First, write your Stan model file. Example:
# tar_stan_example_file() # Writes stantargets_example.stan
# Then in _targets.R, write the pipeline:
targets::tar_pipeline(
  tar_stan_mcmc(
    your_model,
    file = "stantargets_example.stan",
    data = tar_stan_example_data(),
    variables = "beta",
    summaries = list(~quantile(.x, probs = c(0.25, 0.75)))
  )
)

#> <pipeline with 6 targets>