Run a Stan model and return only the summaries.

Not a user-side function. Do not invoke directly.

tar_stan_mle_rep_draws_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  refresh,
  init,
  save_latent_dynamics,
  output_dir,
  algorithm,
  init_alpha,
  iter,
  sig_figs,
  variables
)

Arguments

stan_file
Character, Stan model file or the lines of the file, depending on compile.

stan_name
Friendly suffix of the Stan model target.

stan_path
Original path to the input Stan file.

data
List of data to pass to the Stan model.

compile
Character of length 1. If "original", then cmdstan will compile the source file right before running it (or skip compilation if the binary is up to date). This assumes the worker has access to the file. If the worker is running on a remote computer that does not have access to the model file, set to "copy" instead. compile = "copy" means the pipeline will read the lines of the original Stan model file and send them to the worker. The worker writes the lines to a local copy and compiles the model from there, so it no longer needs access to the original Stan model file on your local machine. However, as a result, the Stan model re-compiles every time the main target reruns.

seed
seed argument to $optimize().

refresh
refresh argument to $optimize().

init
init argument to $optimize().

save_latent_dynamics
save_latent_dynamics argument to $optimize().

output_dir
output_dir argument to $optimize().

algorithm
algorithm argument to $optimize().

init_alpha
init_alpha argument to $optimize().

iter
iter argument to $optimize().

sig_figs
sig_figs argument to $optimize().

variables
variables argument to $draws() and $summary() on the CmdStanMLE object.

Value

A data frame of posterior summaries.