Run a Stan model and return only the summaries.

Not a user-side function. Do not invoke directly.

tar_stan_mle_rep_summary_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  refresh,
  init,
  save_latent_dynamics,
  output_dir,
  algorithm,
  init_alpha,
  iter,
  sig_figs,
  tol_obj,
  tol_rel_obj,
  tol_grad,
  tol_rel_grad,
  tol_param,
  history_size,
  copy_data,
  variables,
  summaries,
  summary_args
)

Arguments

stan_file
The path to a .stan file containing a Stan program. The helper function write_stan_file() is provided for cases when it is more convenient to specify the Stan program as a string.

stan_name
Friendly suffix of the Stan model target.

stan_path
Original path to the input Stan file.

data
(multiple options) The data to use for the variables specified in the data block of the Stan program. One of the following:

A named list of R objects (like for RStan). Internally this list is then written to JSON for CmdStan using write_stan_json().

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan manual for details on using these formats.

NULL or an empty list if the Stan program has no data block.

compile
Character of length 1. If "original", then cmdstan will compile the source file right before running it (or skip compilation if the binary is up to date). This assumes the worker has access to the file. If the worker is running on a remote computer that does not have access to the model file, set to "copy" instead. compile = "copy" means the pipeline will read the lines of the original Stan model file and send them to the worker. The worker writes the lines to a local copy and compiles the model from there, so it no longer needs access to the original Stan model file on your local machine. However, as a result, the Stan model re-compiles every time the main target reruns.

seed
(positive integer) A seed for the (P)RNG to pass to CmdStan.

refresh
(non-negative integer) The number of iterations between printed screen updates. If refresh = 0, only error messages will be printed.

init
(multiple options) The initialization method to use for the variables declared in the parameters block of the Stan program:

A real number x>0. This initializes all parameters randomly between [-x,x] (on the unconstrained parameter space);

The number 0. This initializes all parameters to 0;

A character vector of paths (one per chain) to JSON or Rdump files containing initial values for all or some parameters. See write_stan_json() to write R objects to JSON files compatible with CmdStan.

A list of lists containing initial values for all or some parameters. For MCMC the list should contain a sublist for each chain. For optimization and variational inference there should be just one sublist. The sublists should have named elements corresponding to the parameters for which you are specifying initial values. See Examples.

A function that returns a single list with names corresponding to the parameters for which you are specifying initial values. The function can take no arguments or a single argument chain_id. For MCMC, if the function has argument chain_id it will be supplied with the chain id (from 1 to number of chains) when called to generate the initial values. See Examples.

save_latent_dynamics
(logical) Should auxiliary diagnostic information about the latent dynamics be written to temporary diagnostic CSV files? This argument replaces CmdStan's diagnostic_file argument and the content written to CSV is controlled by the user's CmdStan installation and not CmdStanR (for some algorithms no content may be written). The default is FALSE, which is appropriate for almost every use case. To save the temporary files created when save_latent_dynamics=TRUE see the $save_latent_dynamics_files() method.

output_dir
(string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default), then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()). These temporary files are removed when the fitted model object is garbage collected (manually or automatically).

If a path, then the files are created in output_dir with names corresponding to the defaults used by $save_output_files().

algorithm
(string) The optimization algorithm. One of "lbfgs", "bfgs", or "newton". The control parameters below are only available for "lbfgs" and "bfgs. For their default values and more details see the CmdStan User's Guide. The default values can also be obtained by running cmdstanr_example(method="optimize")$metadata().

init_alpha
(positive real) The initial step size parameter.

iter
(positive integer) The maximum number of iterations.

sig_figs
(positive integer) The number of significant figures used when storing the output values. By default, CmdStan represent the output values with 6 significant figures. The upper limit for sig_figs is 18. Increasing this value will result in larger output CSV files and thus an increased usage of disk space.

tol_obj
(positive real) Convergence tolerance on changes in objective function value.

tol_rel_obj
(positive real) Convergence tolerance on relative changes in objective function value.

tol_grad
(positive real) Convergence tolerance on the norm of the gradient.

tol_rel_grad
(positive real) Convergence tolerance on the relative norm of the gradient.

tol_param
(positive real) Convergence tolerance on changes in parameter value.

history_size
(positive integer) The size of the history used when approximating the Hessian. Only available for L-BFGS.

copy_data
Character vector of names of scalars in data. These values will be inserted as columns in the output data frame for each rep. Useful for simulation studies where you want to check the results against some "true value" in the data. See the mcmc_rep vignette for an example.

Value

A data frame of posterior summaries.