One summary of a CmdStanFit object.

Target to run the $summary() method of a CmdStanFit object.

tar_stan_summary(
  name,
  fit,
  variables = NULL,
  summaries = NULL,
  summary_args = NULL,
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Symbol, base name for the collection of targets. Serves as a prefix for target names.

fit
Symbol, name of a CmdStanFit object or an upstream target that returns a CmdStanFit object.

variables
(character vector) The variables to include.

summaries
Optional list of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
Optional list of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier. Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle()).

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

A targets::tar_target() object to summarize a CmdStanFit object.

Details

tar_stan_mcmc() etc. with summary = TRUE already gives you a target with a $summary() object. Use tar_stan_summary() to create additional specialized summaries.

Examples

# First, write your Stan model file. Example:
# tar_stan_example_file() # Writes stantargets_example.stan
# Then in _targets.R, write the pipeline:
list(
  # Run a model and produce default summaries.
  tar_stan_mcmc(
    your_model,
    stan_files = "stantargets_example.stan",
    data = tar_stan_example_data()
  ),
  # Produce a more specialized summary
  tar_stan_summary(
    your_summary,
    fit = your_model,
    variables = "beta",
    summaries = list(~quantile(.x, probs = c(0.25, 0.75)))
  )
)

#> [[1]]
#> [[1]]$your_model_file_stantargets_example
#> <stem target> 
#>   name: your_model_file_stantargets_example 
#>   command:
#>     "stantargets_example.stan" 
#>   format: file 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: main 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL
#> [[1]]$your_model_mcmc_stantargets_example
#> <stem target> 
#>   name: your_model_mcmc_stantargets_example 
#>   command:
#>     stantargets::tar_stan_mcmc_run(stan_file = your_model_file_stantargets_example, 
#>         data = your_model_data, compile = "original", quiet = TRUE, 
#>         log = NULL, dir = NULL, include_paths = NULL, cpp_options = list(), 
#>         stanc_options = list(), force_recompile = FALSE, seed = NULL, 
#>         refresh = NULL, init = NULL, save_latent_dynamics = FALSE, 
#>         output_dir = NULL, chains = 4, parallel_chains = 1, chain_ids = 1:4, 
#>         threads_per_chain = NULL, iter_warmup = NULL, iter_sampling = NULL, 
#>         save_warmup = FALSE, thin = NULL, max_treedepth = NULL, adapt_engaged = TRUE, 
#>         adapt_delta = NULL, step_size = NULL, metric = NULL, metric_file = NULL, 
#>         inv_metric = NULL, init_buffer = NULL, term_buffer = NULL, 
#>         window = NULL, fixed_param = FALSE, sig_figs = NULL, validate_csv = TRUE, 
#>         show_messages = TRUE, variables = NULL, inc_warmup = FALSE) 
#>   format: qs 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL
#> [[1]]$your_model_draws_stantargets_example
#> <stem target> 
#>   name: your_model_draws_stantargets_example 
#>   command:
#>     tibble::as_tibble(posterior::as_draws_df(your_model_mcmc_stantargets_example$draws(variables = NULL, 
#>         inc_warmup = FALSE))) 
#>   format: fst_tbl 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL
#> [[1]]$your_model_summary_stantargets_example
#> <stem target> 
#>   name: your_model_summary_stantargets_example 
#>   command:
#>     tibble::tibble(your_model_mcmc_stantargets_example$summary(variables = identity(NULL), 
#>         .args = list())) 
#>   format: fst_tbl 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL
#> [[1]]$your_model_diagnostics_stantargets_example
#> <stem target> 
#>   name: your_model_diagnostics_stantargets_example 
#>   command:
#>     tibble::as_tibble(posterior::as_draws_df(your_model_mcmc_stantargets_example$sampler_diagnostics())) 
#>   format: fst_tbl 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL
#> [[1]]$your_model_data
#> <stem target> 
#>   name: your_model_data 
#>   command:
#>     tar_stan_example_data() 
#>   format: qs 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     stantargets
#>     stats
#>     graphics
#>     grDevices
#>     utils
#>     datasets
#>     methods
#>     base 
#>   library:
#>     NULL
#> 
#> [[2]]
#> <stem target> 
#>   name: your_summary 
#>   command:
#>     tibble::tibble(your_model$summary(variables = "beta", 
#>         ~quantile(.x, probs = c(0.25, 0.75)))) 
#>   format: fst_tbl 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     character(0) 
#>   library:
#>     NULL