One variational Bayes run per model with multiple outputs.

Targets to run a Stan model once with variational Bayes and save multiple outputs.

tar_stan_vb(
  name,
  stan_files,
  data = list(),
  compile = c("original", "copy"),
  quiet = TRUE,
  dir = NULL,
  include_paths = NULL,
  cpp_options = list(),
  stanc_options = list(),
  force_recompile = FALSE,
  seed = NULL,
  refresh = NULL,
  init = NULL,
  save_latent_dynamics = FALSE,
  output_dir = NULL,
  algorithm = NULL,
  iter = NULL,
  grad_samples = NULL,
  elbo_samples = NULL,
  eta = NULL,
  adapt_engaged = NULL,
  adapt_iter = NULL,
  tol_rel_obj = NULL,
  eval_elbo = NULL,
  output_samples = NULL,
  sig_figs = NULL,
  variables = NULL,
  summaries = list(),
  summary_args = list(),
  draws = TRUE,
  summary = TRUE,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Symbol, base name for the collection of targets. Serves as a prefix for target names.

stan_files
Character vector of Stan model files. If you supply multiple files, each model will run on the one shared dataset generated by the code in data. If you supply an unnamed vector, tools::file_path_sans_ext(basename(stan_files)) will be used as target name suffixes. If stan_files is a named vector, the suffixed will come from names(stan_files).

data
(multiple options) The data to use for the variables specified in the data block of the Stan program. One of the following:

A named list of R objects (like for RStan). Internally this list is then written to JSON for CmdStan using write_stan_json().

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan manual for details on using these formats.

NULL or an empty list if the Stan program has no data block.

compile
Do compilation? The default is TRUE. If FALSE compilation can be done later via the $compile() method.

quiet
(logical) Should the verbose output from CmdStan during compilation be suppressed? The default is TRUE, but if you encounter an error we recommend trying again with quiet=FALSE to see more of the output.

dir
(string) The path to the directory in which to store the CmdStan executable (or .hpp file if using $save_hpp_file()). The default is the same location as the Stan program.

include_paths
(character vector) Paths to directories where Stan should look for files specified in #include directives in the Stan program.

cpp_options
(list) Any makefile options to be used when compiling the model (STAN_THREADS, STAN_MPI, STAN_OPENCL, etc.). Anything you would otherwise write in the make/local file.

stanc_options
(list) Any Stan-to-C++ transpiler options to be used when compiling the model. See the Examples section below as well as the stanc chapter of the CmdStan Guide for more details on available options: https://mc-stan.org/docs/cmdstan-guide/stanc.html.

force_recompile
(logical) Should the model be recompiled even if was not modified since last compiled. The default is FALSE.

seed
(positive integer) A seed for the (P)RNG to pass to CmdStan.

refresh
(non-negative integer) The number of iterations between printed screen updates. If refresh = 0, only error messages will be printed.

init
(multiple options) The initialization method to use for the variables declared in the parameters block of the Stan program:

A real number x>0. This initializes all parameters randomly between [-x,x] (on the unconstrained parameter space);

The number 0. This initializes all parameters to 0;

A character vector of paths (one per chain) to JSON or Rdump files containing initial values for all or some parameters. See write_stan_json() to write R objects to JSON files compatible with CmdStan.

A list of lists containing initial values for all or some parameters. For MCMC the list should contain a sublist for each chain. For optimization and variational inference there should be just one sublist. The sublists should have named elements corresponding to the parameters for which you are specifying initial values. See Examples.

A function that returns a single list with names corresponding to the parameters for which you are specifying initial values. The function can take no arguments or a single argument chain_id. For MCMC, if the function has argument chain_id it will be supplied with the chain id (from 1 to number of chains) when called to generate the initial values. See Examples.

save_latent_dynamics
(logical) Should auxiliary diagnostic information about the latent dynamics be written to temporary diagnostic CSV files? This argument replaces CmdStan's diagnostic_file argument and the content written to CSV is controlled by the user's CmdStan installation and not CmdStanR (for some algorithms no content may be written). The default is FALSE, which is appropriate for almost every use case. To save the temporary files created when save_latent_dynamics=TRUE see the $save_latent_dynamics_files() method.

output_dir
(string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default), then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()). These temporary files are removed when the fitted model object is garbage collected (manually or automatically).

If a path, then the files are created in output_dir with names corresponding to the defaults used by $save_output_files().

algorithm
(string) The algorithm. Either "meanfield" or "fullrank".

iter
(positive integer) The maximum number of iterations.

grad_samples
(positive integer) The number of samples for Monte Carlo estimate of gradients.

elbo_samples
(positive integer) The number of samples for Monte Carlo estimate of ELBO (objective function).

eta
(positive real) The step size weighting parameter for adaptive step size sequence.

adapt_engaged
(logical) Do warmup adaptation? The default is TRUE. If a precomputed inverse metric is specified via the inv_metric argument (or metric_file) then, if adapt_engaged=TRUE, Stan will use the provided inverse metric just as an initial guess during adaptation. To turn off adaptation when using a precomputed inverse metric set adapt_engaged=FALSE.

adapt_iter
(positive integer) The maximum number of adaptation iterations.

tol_rel_obj
(positive real) Convergence tolerance on the relative norm of the objective.

eval_elbo
(positive integer) Evaluate ELBO every Nth iteration.

output_samples
(positive integer) Number of approximate posterior samples to draw and save.

sig_figs
(positive integer) The number of significant figures used when storing the output values. By default, CmdStan represent the output values with 6 significant figures. The upper limit for sig_figs is 18. Increasing this value will result in larger output CSV files and thus an increased usage of disk space.

variables
(character vector) The variables to include.

summaries
Optional list of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
Optional list of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

draws
Logical, whether to create a target for posterior draws. Saves posterior::as_draws_df(fit$draws()) to a compressed tibble. Convenient, but duplicates storage.

summary
Logical, whether to create a target for fit$summary().

tidy_eval
Logical, whether to enable tidy evaluation when interpreting command and pattern. If TRUE, you can use the "bang-bang" operator !! to programmatically insert the values of global objects.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier. Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle()).

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

tar_stan_vb(name = x, stan_files = "y.stan", ...) returns a list of targets::tar_target() objects:

x_file_y: reproducibly track the Stan model file.

x_lines_y: contents of the Stan model file. Omitted if compile = "original".

x_data: data for the variational Bayes computation.

x_vb_y: CmdStanVB object with all the VB results.

x_draws_y: tidy data frame of VB draws. Omitted if draws = FALSE.

x_summary_y: tidy data frame of VB summaries. Omitted if summary = FALSE. If you supply multiple models, you will get more (model-specific) targets. All the models share the same dataset.

Details

Most of the arguments are passed to the $compile(), $variational(), and $summary() methods of the CmdStanModel class. If you previously compiled the model in an upstream tar_stan_compile() target, then the model should not recompile.

Examples

# First, write your Stan model file. Example:
# tar_stan_example_file() # Writes stantargets_example.stan
# Then in _targets.R, write the pipeline:
targets::tar_pipeline(
  tar_stan_vb(
    your_model,
    stan_files = "stantargets_example.stan",
    data = tar_stan_example_data(),
    variables = "beta",
    summaries = list(~quantile(.x, probs = c(0.25, 0.75)))
  )
)

#> <pipeline with 5 targets>