Run a Stan model and return only the summaries.

Not a user-side function. Do not invoke directly.

tar_stan_gq_rep_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  output,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  copy_data,
  variables,
  summaries,
  summary_args
)

Arguments

stan_file
The path to a .stan file containing a Stan program. The helper function write_stan_file() is provided for cases when it is more convenient to specify the Stan program as a string.

stan_name
Friendly suffix of the Stan model target.

stan_path
Original path to the input Stan file.

data
(multiple options) The data to use for the variables specified in the data block of the Stan program. One of the following:

A named list of R objects (like for RStan). Internally this list is then written to JSON for CmdStan using write_stan_json().

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan manual for details on using these formats.

NULL or an empty list if the Stan program has no data block.

output
Type of output to create, either "summaries", "draws", or "diagnostics".

fitted_params
(multiple options) The parameter draws to use. One of the following:

A CmdStanMCMC or CmdStanVB fitted model object.

A posterior::draws_array (for MCMC) or posterior::draws_matrix (for VB) object returned by CmdStanR's $draws() method.

A character vector of paths to CmdStan CSV output files.

compile
Do compilation? The default is TRUE. If FALSE compilation can be done later via the $compile() method.

quiet
(logical) Should the verbose output from CmdStan during compilation be suppressed? The default is TRUE, but if you encounter an error we recommend trying again with quiet=FALSE to see more of the output.

dir
(string) The path to the directory in which to store the CmdStan executable (or .hpp file if using $save_hpp_file()). The default is the same location as the Stan program.

include_paths
(character vector) Paths to directories where Stan should look for files specified in #include directives in the Stan program.

cpp_options
(list) Any makefile options to be used when compiling the model (STAN_THREADS, STAN_MPI, STAN_OPENCL, etc.). Anything you would otherwise write in the make/local file.

stanc_options
(list) Any Stan-to-C++ transpiler options to be used when compiling the model. See the Examples section below as well as the stanc chapter of the CmdStan Guide for more details on available options: https://mc-stan.org/docs/cmdstan-guide/stanc.html.

force_recompile
(logical) Should the model be recompiled even if was not modified since last compiled. The default is FALSE.

seed
(positive integer) A seed for the (P)RNG to pass to CmdStan.

output_dir
(string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default), then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()). These temporary files are removed when the fitted model object is garbage collected (manually or automatically).

If a path, then the files are created in output_dir with names corresponding to the defaults used by $save_output_files().

sig_figs
(positive integer) The number of significant figures used when storing the output values. By default, CmdStan represent the output values with 6 significant figures. The upper limit for sig_figs is 18. Increasing this value will result in larger output CSV files and thus an increased usage of disk space.

parallel_chains
(positive integer) The maximum number of MCMC chains to run in parallel. If parallel_chains is not specified then the default is to look for the option "mc.cores", which can be set for an entire R session by options(mc.cores=value). If the "mc.cores" option has not been set then the default is 1.

threads_per_chain
(positive integer) If the model was compiled with threading support, the number of threads to use in parallelized sections within an MCMC chain (e.g., when using the Stan functions reduce_sum() or map_rect()). This is in contrast with parallel_chains, which specifies the number of chains to run in parallel. The actual number of CPU cores used use is parallel_chains*threads_per_chain. For an example of using threading see the Stan case study Reduce Sum: A Minimal Example.

copy_data
Character vector of names of scalars in data. These values will be inserted as columns in the output data frame for each rep. Useful for simulation studies where you want to check the results against some "true value" in the data. See the mcmc_rep vignette for an example.

variables
(character vector) The variables to read in. If NULL (the default) then all variables are included.

Value

A data frame of posterior summaries.